% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_files.R
\name{create_r_files}
\alias{create_r_files}
\title{Title}
\usage{
create_r_files(
  r_files_out_path,
  r_files_source_path = system.file("templates", "r_files.csv", package =
    "saros.structure"),
  r_optionals = TRUE,
  r_add_file_scope = TRUE,
  r_prefix_file_scope = "### ",
  r_add_folder_scope_as_README = FALSE,
  numbering_prefix = TRUE,
  r_numbering_prefix = c("none", "max_local", "max_global"),
  r_numbering_inheritance = TRUE,
  word_separator = NULL,
  r_numbering_parent_child_separator = word_separator,
  r_numbering_name_separator = " ",
  r_case = c("asis", "sentence", "title", "lower", "upper", "snake")
)
}
\arguments{
\item{r_files_out_path}{String, path to where to place R placeholder files. If NULL, will not create any.}

\item{r_files_source_path}{String, path to where to find CSV-fiel containing the columns folder_name, folder_scope, file_name, file_scope. If NULL, defaults to system.file("templates", "r_files.csv")).}

\item{r_add_folder_scope_as_README}{}

\item{numbering_prefix}{String. One of c("none", "max_local", "max_global").}

\item{word_separator}{String. Replace separators between words in folder names. Defaults to NULL.}
}
\description{
Title
}
