% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zip_to_folder.R
\name{download_zip_to_folder}
\alias{download_zip_to_folder}
\title{Wrapper to Download and Unzip a Github Repo to A Folder}
\usage{
download_zip_to_folder(
  github_zip_url = "https://github.com/NIFU-NO/nifutemplates/archive/refs/heads/main.zip",
  zip_path = tempfile(fileext = ".zip"),
  files = NULL,
  out_path = getwd(),
  prompt = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{github_zip_url}{URL to zip file, as string.}

\item{zip_path}{String, where to store zip-file. Defaults to a temporary location.}

\item{out_path}{String, directory to where to store the unzipped files.}

\item{overwrite}{Flag, whether to overwrite files in out_path. Defaults to FALSE.}
}
\value{
Character vector of unzipped files.
}
\description{
Wrapper to Download and Unzip a Github Repo to A Folder
}
\examples{
download_zip_to_folder(
   github_zip_url = "https://github.com/NIFU-NO/nifutemplates/archive/refs/heads/main.zip",
   out_path = tempfolder())
}
