% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory_structure.R
\name{handle_numbering_inheritance}
\alias{handle_numbering_inheritance}
\title{Handle numbering inheritance}
\usage{
handle_numbering_inheritance(
  counter = 1,
  numbering_prefix = "none",
  max_folder_count_digits = 0,
  parent_path = "Journal manuscripts",
  parent_numbering = NA,
  numbering_parent_child_separator = "_",
  count_existing_folders = FALSE
)
}
\arguments{
\item{counter}{digit}

\item{numbering_prefix}{One of "none", "0", "00", "000", "max_observed"}

\item{parent_path}{String, path to parent folder}

\item{parent_numbering}{String}

\item{numbering_parent_child_separator}{String}

\item{count_existing_folders}{Boolean}
}
\value{
String
}
\description{
Handle numbering inheritance
}
