% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_folder_structure.R
\name{init_folder_structure}
\alias{init_folder_structure}
\title{Initialize Folder Structure}
\usage{
init_folder_structure(
  path,
  saros_folder = TRUE,
  non_saros_folders = TRUE,
  r_files = TRUE,
  readme_files = TRUE,
  main_folder_numbering = TRUE,
  word_separator = "_"
)
}
\arguments{
\item{path}{String, path to where to create the project files}

\item{saros_folder}{Flag, defaults to TRUE. Whether to include the Saros system.}

\item{non_saros_folders}{Flag, defaults to TRUE. Whether to include all the other folders.}

\item{r_files}{Flag, defaults to TRUE. Whether to include template R files.}

\item{readme_files}{Flag, defaults to TRUE. Whether to include assistance files.}

\item{main_folder_numbering}{Flag, defaults to TRUE. Whether main folders are numbered. Adviced to leave on.}

\item{word_separtor}{String, defaults to "_". to be inserted between all words to ensure valid path names.}
}
\description{
Can be used programatically from the console, or simply use the New Project Wizard..
}
\examples{
init_folder_structure(path = getwd())
}
