% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_folder_structure.R
\name{initialize_saros_project}
\alias{initialize_saros_project}
\title{Initialize Folder Structure}
\usage{
initialize_saros_project(
  path,
  structure_path = NULL,
  numbering_prefix = c("none", "max_local", "max_global"),
  numbering_inheritance = TRUE,
  word_separator = NULL,
  numbering_parent_child_separator = word_separator,
  case = c("asis", "sentence", "title", "lower", "upper", "snake"),
  count_existing_folders = FALSE,
  r_files_out_path = NULL,
  r_files_source_path = system.file("templates", "r_files.csv", package =
    "saros.structure"),
  r_optionals = TRUE,
  r_add_file_scope = TRUE,
  r_prefix_file_scope = "### ",
  r_add_folder_scope_as_README = FALSE
)
}
\arguments{
\item{path}{String, path to where to create the project files}

\item{structure_path}{String. Path to the YAML file that defines the folder structure. Defaults to system.file("templates", "_structure_en.yaml").}

\item{numbering_prefix}{String. One of c("none", "max_local", "max_global").}

\item{numbering_inheritance}{Flag. Whether to inherit numbering from parent folder.}

\item{word_separator}{String. Replace separators between words in folder names. Defaults to NULL.}

\item{numbering_parent_child_separator}{String. Defaults to word_separator.}

\item{case}{String. One of c("asis", "sentence", "lower", "upper", "title", "snake").}

\item{count_existing_folders}{Boolean. Defaults to FALSE.}

\item{r_files_out_path}{String, path to where to place R placeholder files. If NULL, will not create any.}

\item{r_files_source_path}{String, path to where to find CSV-fiel containing the columns folder_name, folder_scope, file_name, file_scope. If NULL, defaults to system.file("templates", "r_files.csv")).}

\item{r_add_folder_scope_as_README}{}

\item{create}{Boolean. Defaults to FALSE.}
}
\description{
Can be used programatically from the console, or simply use the New Project Wizard..
}
\examples{
initialize_saros_project(path = getwd())
}
